backend/requirements.txt
fastapi
starlette
pydantic
pydantic[email]
uvicorn
gunicorn

click
progress
progressbar
maxminddb

coloredlogs
verboselogs
string-color

requests
types-requests

python-jose
pyjwt
passlib
bcrypt

pyotp
pillow
qrcode

pymongo

beautifulsoup4
apscheduler
fake_useragent
tinydb
lxml
pytz
types-pytz
pydub
jinja2
